Add files via upload
Phonebook using C++ is designed to store the contacts details. The Phonebook is an application-based project. It is designed for the people to store the information about their contacts. Rather than going through the pages of their diaries and copies to search an information. So, why do we not keep it in our computer as a record? You may say that we could have used word or notepad but, in that case, it would be a little difficult. The Phonebook Management System provides an easy interface to add or view all or search or modify or delete contact information without going through the pages. A Phonebook Management System will store the information about their contact information such as, name, phone, address, email address, etc.